[chore] move headers into THST folder && add cmake install scripts
