+++ include/THST/predicates.h
